+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Mark `\Wizaplace\Catalog\Product::getUrl` as deprecated
 - Mark `\Wizaplace\Catalog\ProductSummary::getUrl` as deprecated
 - Added `\Wizaplace\Catalog\CatalogService::reportProduct`
+- Added `\Wizaplace\Basket\BasketCompany::getSlug`
 - Added `\Wizaplace\Catalog\CatalogService::getAttributeVariants`
 
 ### Bugfixes